+++ django_elasticsearch_dsl/management/commands/search_index.py
@@ -1,6 +1,8 @@
 from __future__ import unicode_literals, absolute_import
+
+from django.conf import settings
 from django.core.management.base import BaseCommand, CommandError
-from django.utils.six.moves import input
+from six.moves import input
 from ...registries import registry
 
 
@@ -49,6 +51,26 @@ class Command(BaseCommand):
             dest='force',
             help="Force operations without asking"
         )
+        parser.add_argument(
+            '--parallel',
+            action='store_true',
+            dest='parallel',
+            help='Run populate/rebuild update multi threaded'
+        )
+        parser.add_argument(
+            '--no-parallel',
+            action='store_false',
+            dest='parallel',
+            help='Run populate/rebuild update single threaded'
+        )
+        parser.set_defaults(parallel=getattr(settings, 'ELASTICSEARCH_DSL_PARALLEL', False))
+        parser.add_argument(
+            '--no-count',
+            action='store_false',
+            default=True,
+            dest='count',
+            help='Do not include a total count in the summary log line'
+        )
 
     def _get_models(self, args):
         """
@@ -80,30 +102,33 @@ class Command(BaseCommand):
 
     def _create(self, models, options):
         for index in registry.get_indices(models):
-            self.stdout.write("Creating index '{}'".format(index))
+            self.stdout.write("Creating index '{}'".format(index._name))
             index.create()
 
     def _populate(self, models, options):
+        parallel = options['parallel']
         for doc in registry.get_documents(models):
-            qs = doc().get_queryset()
-            self.stdout.write("Indexing {} '{}' objects".format(
-                qs.count(), doc._doc_type.model.__name__)
+            self.stdout.write("Indexing {} '{}' objects {}".format(
+                doc().get_queryset().count() if options['count'] else "all",
+                doc.django.model.__name__,
+                "(parallel)" if parallel else "")
             )
-            doc().update(qs)
+            qs = doc().get_indexing_queryset()
+            doc().update(qs, parallel=parallel)
 
     def _delete(self, models, options):
-        index_names = [str(index) for index in registry.get_indices(models)]
+        index_names = [index._name for index in registry.get_indices(models)]
 
         if not options['force']:
             response = input(
                 "Are you sure you want to delete "
-                "the '{}' indexes? [n/Y]: ".format(", ".join(index_names)))
+                "the '{}' indexes? [y/N]: ".format(", ".join(index_names)))
             if response.lower() != 'y':
                 self.stdout.write('Aborted')
                 return False
 
         for index in registry.get_indices(models):
-            self.stdout.write("Deleting index '{}'".format(index))
+            self.stdout.write("Deleting index '{}'".format(index._name))
             index.delete(ignore=404)
         return True