+++ tests/models.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 from django.db import models
-from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
+from six import python_2_unicode_compatible
 
 
 @python_2_unicode_compatible
@@ -87,3 +87,17 @@ class Ad(models.Model):
 
     def __str__(self):
         return self.title
+
+
+@python_2_unicode_compatible
+class Article(models.Model):
+    slug = models.CharField(
+        max_length=255,
+        unique=True,
+    )
+
+    class Meta:
+        app_label = 'tests'
+
+    def __str__(self):
+        return self.slug