+++ tests/test_documents.py
@@ -1,15 +1,20 @@
+import json
 from unittest import TestCase
 
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
-from elasticsearch_dsl import GeoPoint
-from mock import patch
+from elasticsearch_dsl import GeoPoint, MetaField
+from mock import patch, Mock, PropertyMock
 
 from django_elasticsearch_dsl import fields
-from django_elasticsearch_dsl.documents import DocType
+from django_elasticsearch_dsl.documents import DocType, Document
 from django_elasticsearch_dsl.exceptions import (ModelFieldNotMappedError,
                                                  RedeclaredFieldError)
+from django_elasticsearch_dsl.registries import registry
 from tests import ES_MAJOR_VERSION
+
+from .models import Article
+from .documents import ArticleDocument, ArticleWithSlugAsIdDocument
 
 
 class Car(models.Model):
@@ -34,56 +39,66 @@ class Manufacturer(models.Model):
         app_label = 'car'
 
 
+@registry.register_document
 class CarDocument(DocType):
     color = fields.TextField()
-    type = fields.StringField()
+    type = fields.TextField()
 
     def prepare_color(self, instance):
         return "blue"
 
     class Meta:
+        doc_type = 'car_document'
+
+    class Django:
         fields = ['name', 'price']
         model = Car
-        index = 'car_index'
         related_models = [Manufacturer]
+
+    class Index:
+        name = 'car_index'
         doc_type = 'car_document'
 
 
 class DocTypeTestCase(TestCase):
 
     def test_model_class_added(self):
-        self.assertEqual(CarDocument._doc_type.model, Car)
+        self.assertEqual(CarDocument.django.model, Car)
 
     def test_ignore_signal_default(self):
-        self.assertFalse(CarDocument._doc_type.ignore_signals)
+        self.assertFalse(CarDocument.django.ignore_signals)
 
     def test_auto_refresh_default(self):
-        self.assertTrue(CarDocument._doc_type.auto_refresh)
+        self.assertTrue(CarDocument.django.auto_refresh)
 
     def test_ignore_signal_added(self):
+
+        @registry.register_document
         class CarDocument2(DocType):
-            class Meta:
+            class Django:
                 model = Car
                 ignore_signals = True
 
-        self.assertTrue(CarDocument2._doc_type.ignore_signals)
+        self.assertTrue(CarDocument2.django.ignore_signals)
 
     def test_auto_refresh_added(self):
+        @registry.register_document
         class CarDocument2(DocType):
-            class Meta:
+            class Django:
                 model = Car
                 auto_refresh = False
 
-        self.assertFalse(CarDocument2._doc_type.auto_refresh)
+        self.assertFalse(CarDocument2.django.auto_refresh)
 
     def test_queryset_pagination_added(self):
+        @registry.register_document
         class CarDocument2(DocType):
-            class Meta:
+            class Django:
                 model = Car
                 queryset_pagination = 120
 
-        self.assertIsNone(CarDocument._doc_type.queryset_pagination)
-        self.assertEqual(CarDocument2._doc_type.queryset_pagination, 120)
+        self.assertIsNone(CarDocument.django.queryset_pagination)
+        self.assertEqual(CarDocument2.django.queryset_pagination, 120)
 
     def test_fields_populated(self):
         mapping = CarDocument._doc_type.mapping
@@ -93,16 +108,17 @@ class DocTypeTestCase(TestCase):
         )
 
     def test_related_models_added(self):
-        related_models = CarDocument._doc_type.related_models
+        related_models = CarDocument.django.related_models
         self.assertEqual([Manufacturer], related_models)
 
     def test_duplicate_field_names_not_allowed(self):
         with self.assertRaises(RedeclaredFieldError):
+            @registry.register_document
             class CarDocument(DocType):
-                color = fields.StringField()
-                name = fields.StringField()
+                color = fields.TextField()
+                name = fields.TextField()
 
-                class Meta:
+                class Django:
                     fields = ['name']
                     model = Car
 
@@ -122,7 +138,6 @@ class DocTypeTestCase(TestCase):
 
         self.assertEqual(
             CarDocument._doc_type.mapping.to_dict(), {
-                'car_document': {
                     'properties': {
                         'name': {
                             'type': text_type
@@ -137,7 +152,6 @@ class DocTypeTestCase(TestCase):
                             'type': 'double'
                         }
                     }
-                }
             }
         )
 
@@ -160,13 +174,16 @@ class DocTypeTestCase(TestCase):
         )
 
     def test_prepare_ignore_dsl_base_field(self):
+        @registry.register_document
         class CarDocumentDSlBaseField(DocType):
             position = GeoPoint()
 
-            class Meta:
+            class Django:
                 model = Car
-                index = 'car_index'
                 fields = ['name', 'price']
+
+            class Index:
+                name = 'car_index'
 
         car = Car(name="Type 57", price=5400000.0, not_indexed="not_indexex")
         doc = CarDocumentDSlBaseField()
@@ -194,7 +211,6 @@ class DocTypeTestCase(TestCase):
                     'color': doc.prepare_color(None),
                 },
                 '_index': 'car_index',
-                '_type': 'car_document'
             }]
             self.assertEqual(1, mock.call_count)
             self.assertEqual(
@@ -202,7 +218,7 @@ class DocTypeTestCase(TestCase):
             )
             self.assertTrue(mock.call_args_list[0][1]['refresh'])
             self.assertEqual(
-                doc.connection, mock.call_args_list[0][1]['client']
+                doc._index.connection, mock.call_args_list[0][1]['client']
             )
 
     def test_model_instance_iterable_update(self):
@@ -223,7 +239,6 @@ class DocTypeTestCase(TestCase):
                     'color': doc.prepare_color(None),
                 },
                 '_index': 'car_index',
-                '_type': 'car_document'
             },
                 {
                     '_id': car2.pk,
@@ -234,8 +249,7 @@ class DocTypeTestCase(TestCase):
                         'type': car2.type(),
                         'color': doc.prepare_color(None),
                     },
-                    '_index': 'car_index',
-                    '_type': 'car_document'
+                    '_index': 'car_index'
                 }]
             self.assertEqual(1, mock.call_count)
             self.assertEqual(
@@ -243,12 +257,12 @@ class DocTypeTestCase(TestCase):
             )
             self.assertTrue(mock.call_args_list[0][1]['refresh'])
             self.assertEqual(
-                doc.connection, mock.call_args_list[0][1]['client']
+                doc._index.connection, mock.call_args_list[0][1]['client']
             )
 
     def test_model_instance_update_no_refresh(self):
         doc = CarDocument()
-        doc._doc_type.auto_refresh = False
+        doc.django.auto_refresh = False
         car = Car()
         with patch('django_elasticsearch_dsl.documents.bulk') as mock:
             doc.update(car)
@@ -256,7 +270,7 @@ class DocTypeTestCase(TestCase):
 
     def test_model_instance_iterable_update_with_pagination(self):
         class CarDocument2(DocType):
-            class Meta:
+            class Django:
                 model = Car
                 queryset_pagination = 2
 
@@ -264,8 +278,136 @@ class DocTypeTestCase(TestCase):
         car1 = Car()
         car2 = Car()
         car3 = Car()
-        with patch('django_elasticsearch_dsl.documents.bulk') as mock:
-            doc.update([car1, car2, car3])
+
+        bulk = "django_elasticsearch_dsl.documents.bulk"
+        parallel_bulk = "django_elasticsearch_dsl.documents.parallel_bulk"
+        with patch(bulk) as mock_bulk, patch(parallel_bulk) as mock_parallel_bulk:
+            doc.update([car3, car2, car3])
             self.assertEqual(
-                3, len(list(mock.call_args_list[0][1]['actions']))
+                3, len(list(mock_bulk.call_args_list[0][1]['actions']))
             )
+            self.assertEqual(mock_bulk.call_count, 1, "bulk is called")
+            self.assertEqual(mock_parallel_bulk.call_count, 0, "parallel bulk is not called")
+
+    def test_model_instance_iterable_update_with_parallel(self):
+        class CarDocument2(DocType):
+            class Django:
+                model = Car
+
+        doc = CarDocument()
+        car1 = Car()
+        car2 = Car()
+        car3 = Car()
+        bulk = "django_elasticsearch_dsl.documents.bulk"
+        parallel_bulk = "django_elasticsearch_dsl.documents.parallel_bulk"
+        with patch(bulk) as mock_bulk, patch(parallel_bulk) as mock_parallel_bulk:
+            doc.update([car1, car2, car3], parallel=True)
+            self.assertEqual(mock_bulk.call_count, 0, "bulk is not called")
+            self.assertEqual(mock_parallel_bulk.call_count, 1, "parallel bulk is called")
+
+    def test_init_prepare_correct(self):
+        """Does init_prepare() run and collect the right preparation functions?"""
+
+        d = CarDocument()
+        self.assertEqual(len(d._prepared_fields), 4)
+
+        expect = {
+            'color': ("<class 'django_elasticsearch_dsl.fields.TextField'>",
+                    ("<class 'method'>", "<type 'instancemethod'>")), # py3, py2
+            'type': ("<class 'django_elasticsearch_dsl.fields.TextField'>",
+                    ("<class 'functools.partial'>","<type 'functools.partial'>")),
+            'name': ("<class 'django_elasticsearch_dsl.fields.TextField'>",
+                    ("<class 'functools.partial'>","<type 'functools.partial'>")),
+            'price': ("<class 'django_elasticsearch_dsl.fields.DoubleField'>",
+                    ("<class 'functools.partial'>","<type 'functools.partial'>")),
+        }
+
+        for name, field, prep in d._prepared_fields:
+            e = expect[name]
+            self.assertEqual(str(type(field)), e[0], 'field type should be copied over')
+            self.assertTrue('__call__' in dir(prep), 'prep function should be callable')
+            self.assertTrue(str(type(prep)) in e[1], 'prep function is correct partial or method')
+
+    def test_init_prepare_results(self):
+        """Are the results from init_prepare() actually used in prepare()?"""
+        d = CarDocument()
+
+        car = Car()
+        setattr(car, 'name', "Tusla")
+        setattr(car, 'price', 340123.21)
+        setattr(car, 'color', "polka-dots") # Overwritten by prepare function
+        setattr(car, 'pk', 4701) # Ignored, not in document
+        setattr(car, 'type', "imaginary")
+
+        self.assertEqual(d.prepare(car), {'color': 'blue', 'type': 'imaginary', 'name': 'Tusla', 'price': 340123.21})
+
+        m = Mock()
+        # This will blow up should we access _fields and try to iterate over it.
+        # Since init_prepare compiles a list of prepare functions, while
+        # preparing no access to _fields should happen
+        with patch.object(CarDocument, '_fields', 33):
+            d.prepare(m)
+        self.assertEqual(sorted([tuple(x) for x in m.method_calls], key=lambda _: _[0]),
+                         [('name', (), {}),  ('price', (), {}), ('type', (), {})]
+        )
+
+    # Mock the elasticsearch connection because we need to execute the bulk so that the generator
+    # got iterated and generate_id called.
+    # If we mock the bulk in django_elasticsearch_dsl.document
+    # the actual bulk will be never called and the test will fail
+    @patch('elasticsearch_dsl.connections.Elasticsearch.bulk')
+    def test_default_generate_id_is_called(self, _):
+        article = Article(
+            id=124594,
+            slug='some-article',
+        )
+        @registry.register_document
+        class ArticleDocument(DocType):
+            class Django:
+                model = Article
+                fields = [
+                    'slug',
+                ]
+
+            class Index:
+                name = 'test_articles'
+                settings = {
+                    'number_of_shards': 1,
+                    'number_of_replicas': 0,
+                }
+
+        with patch.object(ArticleDocument, 'generate_id',
+                          return_value=article.id) as patched_method:
+            d = ArticleDocument()
+            d.update(article)
+            patched_method.assert_called()
+
+    @patch('elasticsearch_dsl.connections.Elasticsearch.bulk')
+    def test_custom_generate_id_is_called(self, mock_bulk):
+        article = Article(
+            id=54218,
+            slug='some-article-2',
+        )
+
+        @registry.register_document
+        class ArticleDocument(DocType):
+            class Django:
+                model = Article
+                fields = [
+                    'slug',
+                ]
+
+            class Index:
+                name = 'test_articles'
+
+            @classmethod
+            def generate_id(cls, article):
+                return article.slug
+
+        d = ArticleDocument()
+        d.update(article)
+
+        # Get the data from the elasticsearch low level API because
+        # The generator get executed there.
+        data = json.loads(mock_bulk.call_args[0][0].split("\n")[0])
+        assert data["index"]["_id"] == article.slug