+++ HISTORY.rst
@@ -2,6 +2,64 @@
 
 History
 -------
+7.1.4 (2020-07-05)
+~~~~~~~~~~~~~~~~~~
+* Configure Elasticsearch _id dynamically from document (#272)
+* Use chain.from_iterable in for performance improvement (#278)
+* Handle case where SimpleLazyObject being treated as an Iterable (#255)
+* Camelcase default value in management command (#254)
+* Various updates and fixup in docs (#250, #276)
+* Start testing against Python 3.8 (#266)
+
+
+7.1.1 (2019-12-26)
+~~~~~~~~~~~~~~~~~~
+* Adding detailed documentation and published to Read The Docs #222
+* Resolve name resolution while delete, create index (#228)
+* Added support for Django 3.0. (#230)
+* Removing old Elasticsearc compatibility (#219)
+* Drop StringField in favor of TextField.
+
+
+7.1.0 (2019-10-29)
+~~~~~~~~~~~~~~~~~~
+* Support for Django `DecimalField` #141
+* Indexing speedup by using `parallel` indexing. #213.
+  Now you can pass `--parallel` or set `ELASTICSEARCH_DSL_PARALLEL`
+  in your settings to get indexing speed boost while indexing
+  through management command.
+* Fixing name resolution in management command #206
+* Small documentation fixes. #196
+
+
+7.0.0 (2019-08-11)
+~~~~~~~~~~~~~~~~~~
+* Support Elasticsearch 7.0 (See PR #176)
+* Added order by to paginate queryset properly (See PR #153)
+* Remove `standard` token filter from `README.md` and test files
+* Various documentation fixes
+
+
+6.4.2 (2019-07-26)
+~~~~~~~~~~~~~~~~~~
+* Fix document importing path
+* Update readme
+
+
+
+6.4.1 (2019-06-14)
+~~~~~~~~~~~~~~~~~~
+* The `DocType` import has changed to `Document`
+
+
+
+6.4.0 (2019-06-01)
+~~~~~~~~~~~~~~~~~~
+* Support elasticsearch-dsl>6.3.0
+* Class `Meta` has changed to class `Django` (See PR #136)
+* Add `register_document` decorator to register a document (See PR #136)
+* Additional Bug fixing and others
+
 
 0.5.1 (2018-11-07)
 ~~~~~~~~~~~~~~~~~~